+++ src/test_libFLAC++/encoders.cpp
@@ -453,7 +453,7 @@ static bool test_stream_encoder(Layer layer, bool is_ogg)
 
 	printf("testing get_total_samples_estimate()... ");
 	if(encoder->get_total_samples_estimate() != streaminfo_.data.stream_info.total_samples) {
-		printf("FAILED, expected %llu, got %llu\n", streaminfo_.data.stream_info.total_samples, encoder->get_total_samples_estimate());
+		printf("FAILED, expected %llu, got %llu\n", (unsigned long long)streaminfo_.data.stream_info.total_samples, (unsigned long long)encoder->get_total_samples_estimate());
 		return false;
 	}
 	printf("OK\n");
